Allow PHP 8.5 support

+++ src/Client/LicenseHistory.php
@@ -294,7 +294,7 @@ class LicenseHistory
     public function getTotalLicenseHistoryPages() : int
     {
         if (!$this->_initial_invalid_state && $this->_current_response->getNbResults() !== null) {
-            return (integer) ceil((double) $this->_current_response->getNbResults() / $this->_current_request->getSearchParams()->getLimit());
+            return (int) ceil((float) $this->_current_response->getNbResults() / $this->_current_request->getSearchParams()->getLimit());
         }
 
         return self::LICENSE_HISTORY_RETURN_ERROR;
@@ -321,7 +321,7 @@ class LicenseHistory
     {
         if (!$this->_initial_invalid_state && $this->_current_response->getNbResults() !== null) {
             $offset_value = $this->_current_request->getSearchParams()->getOffset();
-            $result = (integer) (ceil((double) $offset_value / $this->_current_request->getSearchParams()->getLimit()));
+            $result = (int) (ceil((float) $offset_value / $this->_current_request->getSearchParams()->getLimit()));
             return $result;
         }
